+++ src/app/screens/BoardsList/components/BoardCard/styles.css
